requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
slack-bolt>=1.18.0
certifi>=2024.0.0
